secrets as env
This will make working with Hugging Face tokens easier. This would allow
you to create a secret with the following content:
```
HUGGING_FACE_HUB_TOKEN: my-token
```

Afterwards all Substratus resources will be able to use this
HuggingFace token. This also removes the need for custom entrypoints in
the images that strip PARAM_ from the env variable.